Use a QueryContext for try_mark_green.
